feat: Delete Product

+++ src/BoundedContext/Product/Infrastructure/DeleteProductController.php
@@ -6,6 +6,7 @@ namespace Src\BoundedContext\Product\Infrastructure;
 
 use Illuminate\Http\Request;
 use Src\BoundedContext\Product\Application\DeleteProductUseCase;
+use Src\BoundedContext\Product\Application\GetProductUseCase;
 use Src\BoundedContext\Product\Infrastructure\Eloquent\EloquentProductRepository;
 
 final class DeleteProductController
@@ -21,7 +22,16 @@ final class DeleteProductController
     {
         $productId = (int)$request->id;
 
+        $getProductUseCase = new GetProductUseCase($this->repository);
+        $product           = $getProductUseCase->__invoke($productId);
+
+        if (!$product) {
+            return null;
+        }
+
         $deleteProductUseCase = new DeleteProductUseCase($this->repository);
         $deleteProductUseCase->__invoke($productId);
+
+        return true;
     }
 }
